Adds a source duplication feature
The occam.Source function can be used to create a copy of the given
directory. It also introduces a .occam-key file into that copy that
includes unique content. This means that tests that use this function
and also refer to the same original source code can run in parallel
without concerns that pack with create duplicate images that will cause
test running issues.