Support testing when Docker is not running locally
With the recent changes to Docker licensing, a user may not be running Docker on localhost. The occam code assumes that a test application will be accessible on `localhost` though. If you have a set up where Docker is running in a custom VM or even remotely. The occam matchers will always fail because nothing is available on localhost.

This PR makes adjustments so that you can configure a remote host. It will look at and parse `DOCKER_HOST`. If not set, then it defaults to `localhost` which preserves the current behavior. If set, it will parse the URL and return the hostname so long as it's not a `unix://` socket.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>