feat: allow occam to cope with pack limitations
pack seems to be very limited in the options it will
accept when building for multi-arch. It does not
seem to accept a path with a tgz as input and only
works properly if you build in an exploded directory.
Otherwise it appears to have built the cnb file but
the structure of the bin directory is incorrect and
the executables cannot be found.

Update the jam packager to work around this by exploding
the tgz after creating it and then build in that
exploded directory.

It is still a good idea to build the tgz as that
filters the files based on what is specified in the
buildpack.toml.

Signed-off-by: Michael Dawson <mdawson@devrus.com>