fix: restore ability to use occam with extensions
Recent change to support multi-arch for buildpacks brokle occam
for extensions. This was because extensions were being built
with "pack buildpack package" which seems to work for extensions
if a tgz is passed in but not an expanded directory.

Update to use "pack extension package" for extensions instead.
Unfortunately pack does not yet seem to support multi-arch
for extensions so we'll have to add the --target option
for extensions after that has been added.

Signed-off-by: Michael Dawson <mdawson@devrus.com>